backend/requirements.txt
# Core Dependencies
google-generativeai>=0.3.0
PyYAML>=6.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0

# Development Dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
isort>=5.0.0
mypy>=0.900
pylint>=2.12.0

# Documentation
mkdocs>=1.3.0
mkdocs-material>=8.0.0
